# Geostationary projection: speed up inverse spherical and ellipsoidal computation