Fixes #4522

* Spherical case:
Before:
```
$ bin/bench_proj_trans --pipeline "+proj=pipeline +inv +step +proj=geos +h=35785831.0 +lon_0=0 +ellps=sphere" 137881.85 4490747.09
137881.85 4490747.09 -> 0.034807228056148 0.850471730911161
Duration: 677 ms
Throughput: 7.39 million coordinates/s
```

After:
```
$ bin/bench_proj_trans --pipeline "+proj=pipeline +inv +step +proj=geos +h=35785831.0 +lon_0=0 +ellps=sphere" 137881.85 4490747.09
137881.85 4490747.09 -> 0.034807228056148 0.850471730911161
Duration: 602 ms
Throughput: 8.31 million coordinates/s
```

* Ellipsoidal case:

Before:
```
$ bin/bench_proj_trans --pipeline "+proj=pipeline +inv +step +proj=geos +h=35785831.0 +lon_0=0 +ellps=WGS84" 137881.85 4490747.09
137881.85 4490747.09 -> 0.0349065842446029 0.855211334383648
Duration: 978 ms
Throughput: 5.11 million coordinates/s
```

After:
```
$ bin/bench_proj_trans --pipeline "+proj=pipeline +inv +step +proj=geos +h=35785831.0 +lon_0=0 +ellps=WGS84" 137881.85 4490747.09
137881.85 4490747.09 -> 0.0349065842446029 0.855211334383648
Duration: 668 ms
Throughput: 7.49 million coordinates/s
```